fix rank weight float issue

+++ sql/V220608_cvs.sql
@@ -95,7 +95,7 @@ CREATE TABLE IF NOT EXISTS `seddb`.`cvs_stakeholder_needs`
     `vcs_row`           INT UNSIGNED NOT NULL,
     `need`              TEXT NOT NULL,
     `value_dimension`   TEXT NULL,
-    `rank_weight`       FLOAT NULL,
+    `rank_weight`       DOUBLE(10,5) NULL,
     FOREIGN KEY(`vcs_row`)
         REFERENCES  `seddb`.`cvs_vcs_rows`(`id`)
         ON DELETE CASCADE,
